+++ cob/cli/docker_cli.py
@@ -3,6 +3,7 @@ import click
 import flask_migrate
 import logbook
 import multiprocessing
+from pathlib import Path
 import random
 import shutil
 import string
@@ -46,8 +47,8 @@ def docker():
     pass
 
 
-@docker.command()
-def generate():
+@docker.command(name="generate-docker-file")
+def generate_dockerfile():
     proj = get_project()
     template = load_template('Dockerfile')
 
@@ -94,7 +95,7 @@ def docker_build(sudo, extra_build_args="", use_exec=True, image_name=None, rele
     project = get_project()
     if image_name is None:
         image_name = f'{project.name}:{"latest" if release else "dev"}'.format(project.name, 'latest' if release else 'dev')
-    generate.callback()
+    generate_dockerfile.callback()
     cmd = docker_cmd.build(['-t', image_name, '-f', '.Dockerfile', '.', *extra_build_args.split()]).force_sudo(sudo)
     _logger.debug('Running Command: {}', cmd)
     cmd.run(use_exec=use_exec)
@@ -218,16 +219,16 @@ def _exec_docker_compose(cmd, **kwargs):
     cmd.execv()
 
 
-@docker.command()
+@docker.command(name='generate-docker-compose-file')
 @click.option('--image-name', default=None)
-def compose(image_name):
-    print(_generate_compose_file_string(_generate_compose_file_dict(image_name=image_name)))
+@click.option('--force-config-override', is_flag=True, default=False)
+def generate_docker_compose_file(image_name, force_config_override):
+    """Prints out a docker-compose.yml for this project"""
+    print(_generate_compose_file_string(image_name=image_name, force_config_override=force_config_override))
 
 
-def _generate_compose_file_dict(*, http_port=None, image_name=None):
+def _generate_compose_file_dict(*, http_port=None, image_name=None, force_config_override=False):
     project = get_project()
-
-    local_override_path = get_etc_config_path(project.name)
 
     if image_name is None:
         image_name = f'{project.name}:dev'
@@ -296,15 +297,16 @@ def _generate_compose_file_dict(*, http_port=None, image_name=None):
             'image': 'redis',
         }
 
-    if local_override_path.is_dir():
+    config_override_path = get_etc_config_path(project.name)
+    if force_config_override or config_override_path.is_dir():
         for service_config in services.values():
-            service_config.setdefault('volumes', []).append('{0}:{0}'.format(local_override_path))
+            service_config.setdefault('volumes', []).append('{0}:{0}'.format(config_override_path))
 
     return config
 
 
-def _generate_compose_file_string(*args, **kwargs):
-    config = _generate_compose_file_dict(*args, **kwargs)
+def _generate_compose_file_string(**kwargs):
+    config = _generate_compose_file_dict(**kwargs)
     return _dump_yaml(config)
 
 
@@ -340,3 +342,18 @@ def test(build_image, sudo):
         'bash', '-c', "rsync -rvP --delete --exclude .cob /localdir/ /app/ && cob test"])
     if cmd.popen().wait() != 0:
         raise TestsFailed('Tests failed')
+
+
+@docker.command(name="run-image", help='Runs a cob project in a pre-built docker image')
+@click.argument('image_name')
+def run_image(image_name):
+    project = get_project()
+    cmd = docker_cmd.run(['--rm', image_name, 'cob', 'docker', 'generate-docker-compose-file', '--image-name', image_name])
+    if get_etc_config_path(project.name).is_dir():
+        cmd.args(['--force-config-override'])
+    compose_file_contents = cmd.check_output()
+    compose_path = Path('/tmp') / f"__cob_docker_compose_{image_name.replace(':', '__')}.yml"
+    with compose_path.open('wb') as f:
+        f.write(compose_file_contents)
+    cmd = docker_compose_cmd.args(['-f', compose_path, 'up'])
+    cmd.execv()