+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`82` Added `cob docker run-image` to run a prebuilt cob image without requiring dependencies
 * :feature:`77` Cob now required Python 3.6
 * :feature:`50` ``cob docker test`` can now be used to run your tests inside a working
   docker-compose setup